Crate next_core

Modules

env

manifest

next_client

next_config

next_image

next_server

next_shared

react_refresh

source_map

Functions

create_app_source

Create a content source serving the app or src/app directory as Next.js app folder.

create_page_source

Create a content source serving the pages or src/pages directory as Next.js pages folder.

create_web_entry_source

register