Module next_core::app_structure

Structs

AppStructure

A (sub)directory in the app directory with all analyzed routes and folders.

AppStructureItemVc

Vc for AppStructureItem

AppStructureVc

Vc for AppStructure

OptionAppStructure

OptionAppStructureVc

Vc for OptionAppStructure

Enums

AppStructureItem

A final route in the app directory.

Functions

find_app_structure

Finds and returns the AppStructure of the app directory if enabled and existing.

find_app_structure_inline

Finds and returns the AppStructure of the app directory if enabled and existing.

get_app_structure

Parses a directory as app directory and returns the AppStructure.

get_app_structure_inline

Parses a directory as app directory and returns the AppStructure.

Type Definitions

AppStructureItemReadRef

see [turbo_tasks::ReadRef]

AppStructureReadRef

see [turbo_tasks::ReadRef]

OptionAppStructureReadRef

see [turbo_tasks::ReadRef]