Module next_core::env

Functions

env_for_js

Creates a ProcessEnvVc safe to use in JS, by stringifying and encoding as regular JS strings. Setting client to true will additionally filter the env to just the keys that are acceptable for the client to access.

env_for_js_inline

Creates a ProcessEnvVc safe to use in JS, by stringifying and encoding as regular JS strings. Setting client to true will additionally filter the env to just the keys that are acceptable for the client to access.

load_env

Loads a series of dotenv files according to the precedence rules set by https://nextjs.org/docs/basic-features/environment-variables#environment-variable-load-order

load_env_inline

Loads a series of dotenv files according to the precedence rules set by https://nextjs.org/docs/basic-features/environment-variables#environment-variable-load-order