Module next_core::pages_structure

Structs

OptionPagesStructure

OptionPagesStructureVc

Vc for OptionPagesStructure

PagesStructure

A (sub)directory in the pages directory with all analyzed routes and folders.

PagesStructureItemVc

Vc for PagesStructureItem

PagesStructureVc

Vc for PagesStructure

Enums

PagesStructureItem

A final route in the pages directory.

Functions

find_pages_structure

Finds and returns the PagesStructure of the pages directory if existing.

find_pages_structure_inline

Finds and returns the PagesStructure of the pages directory if existing.

get_pages_structure

Parses a directory as pages directory and returns the PagesStructure.

get_pages_structure_inline

Parses a directory as pages directory and returns the PagesStructure.

Type Definitions

OptionPagesStructureReadRef

see [turbo_tasks::ReadRef]

PagesStructureItemReadRef

see [turbo_tasks::ReadRef]

PagesStructureReadRef

see [turbo_tasks::ReadRef]