Module next_core::react_refresh

Structs

AssertReactRefreshResultVc

Vc for AssertReactRefreshResult

ReactRefreshResolvingIssue

An issue that occurred while resolving the React Refresh runtime module.

ReactRefreshResolvingIssueVc

Vc for ReactRefreshResolvingIssue

Enums

AssertReactRefreshResult

Functions

assert_can_resolve_react_refresh

Checks whether we can resolve the React Refresh runtime module from the given path. Emits an issue if we can’t.

assert_can_resolve_react_refresh_inline

Checks whether we can resolve the React Refresh runtime module from the given path. Emits an issue if we can’t.

resolve_react_refresh

Resolves the React Refresh runtime module from the given [AssetContextVc].

resolve_react_refresh_inline

Resolves the React Refresh runtime module from the given [AssetContextVc].

Type Definitions

AssertReactRefreshResultReadRef

see [turbo_tasks::ReadRef]

ReactRefreshResolvingIssueReadRef

see [turbo_tasks::ReadRef]